+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elf2uf2-rs"
-version = "1.3.7"
+version = "1.4.0-dev"
 authors = ["Jonathan Nilsson <jonathan@voysys.se>"]
 edition = "2018"
 
@@ -10,12 +10,21 @@ readme = "README.md"
 license = "0BSD"
 repository = "https://github.com/JoNil/elf2uf2-rs"
 
+[[bin]]
+name = "elf2uf2-rs"
+required-features = ["binary"]
+
+[features]
+default = ["binary"]
+binary = ["clap", "once_cell", "progress_bar", "serialport", "sysinfo"]
+progress_bar = ["pbr"]
+
 [dependencies]
-assert_into = "1.1"
-clap = { version = "3.0.0-rc.4", features = ["derive"] }
-once_cell = "1.5"
-pbr = "1"
-serialport = "4"
-static_assertions = "1"
-sysinfo = "0.20"
+assert_into = { version = "1.1" }
+clap = { version = "3.0.0-rc.4", features = ["derive"], optional = true }
+once_cell = { version = "1.5", optional = true }
+pbr = { version = "1", optional = true }
+serialport = { version = "4", optional = true }
+static_assertions = { version = "1" }
+sysinfo = { version = "0.20", optional = true }
 zerocopy = "0.6"
